All the animals work!
Most I think.  The tigers mat br dropping through 0,0.

+++ Assets/Scripts/PlayerDataSO.cs
@@ -7,8 +7,8 @@ using UnityEngine;
 public class PlayerDataSO : ScriptableObject
 {
     [SerializeField]
-    private float _gameWave;
-    public float GameWave
+    private int _gameWave;
+    public int GameWave
     {
         get
         {
